# Conflicts:
#	vehicle/OVMS.V3/components/vehicle_renaulttwizy/src/vehicle_renaulttwizy.cpp

+++ vehicle/OVMS.V3/main/ovms_utils.h
@@ -121,8 +121,9 @@ char* HexByte(char* p, uint8_t byte);
 /**
  * FormatHexDump: create/fill hexdump buffer including printable representation
  * Note: allocates buffer as necessary in *bufferp, caller must free.
+ * Returns new remaining length
  */
-int FormatHexDump(char** bufferp, const char* data, size_t rlength, size_t colsize=16);
+size_t FormatHexDump(char** bufferp, const char* data, size_t rlength, size_t colsize=16);
 
 
 /**
